+++ src/configs/updatedList.ts
@@ -1,0 +1,10 @@
+// Next.js 15 - src/configs/updatedList.ts
+export const updatedList = [
+    'added: Update Version Dialog',
+    'updated: Skill Description (add parameter "Ranged Block")',
+    'updated: Skill Description (add parameter "Element Attack")',
+    'updated: Skill Description (add parameter "Sword Attack")',
+    'updated: Skill Description (add parameter "Axe Attack")',
+    'updated: Skill Description (add parameter "Critical Resist")',
+    'updated: Skill Description (add parameter "Remove All Debuff")'
+];